improved functionality, fixed a lot of bugs.
Fixed a view adjusting problem that didn’t look good.
Fixed a bug that was causing app to crash after QR code scan
Fixed problem of list of add new currency icons disappearing on screen
rotate.
Restructured the app to include the OWCoin.java class and the
WalletManager class which both seem like they are going to be very
helpful.
Mainly just improved the functionality of the app, fixed a lot of bugs.